refactor: make loader as submodule
